pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ygo-effect-dsl"
version = "0.1.0"
description = "CORE: ingest JSONL (etl output) -> transform to DSL YAML -> analyze -> optional IR compile"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
ygo-effect-dsl = "ygo_effect_dsl.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
